+++ src/traits.rs
@@ -7,7 +7,7 @@ use core::{
 };
 use ff::PrimeField;
 use merlin::Transcript;
-use rug::Integer;
+use num_bigint::BigInt;
 
 /// Represents an element of a group
 pub trait Group:
@@ -50,7 +50,7 @@ pub trait Group:
   fn to_coordinates(&self) -> (Self::Base, Self::Base, bool);
 
   /// Returns the order of the group as a big integer
-  fn get_order() -> Integer;
+  fn get_order() -> BigInt;
 }
 
 /// Represents a compressed version of a group element
